pyproject.toml
[tool.poetry]
name = "abc"
version = "0.1.0"
description = "Test assignment for Kvint"
authors = ["ptjexio@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"
aio-pika = "^8.2.4"
dependency-injector = "^4.40.0"
python-dotenv = "^0.21.0"
pyyaml = "^6.0"
aiogram = "3.0.0b5"


[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
yesqa = "^1.4.0"
pre-commit = "^2.20.0"
autoflake = "^1.7.7"
black = "^22.10.0"
commitizen = "^2.37.0"
invoke = "^1.7.3"
safety = "^2.3.1"
interrogate = "^1.5.0"
flake8-class-attributes-order = "^0.1.3"
flake8-annotations = "^2.9.1"
isort = "^5.10.1"
perflint = "^0.7.3"
flakeheaven = "^3.2.0"
darglint = "^1.8.1"
wemake-python-styleguide = "^0.17.0"
aiogram-tests = "^1.0.3"
pytest = "^7.2.0"

[tool.flakeheaven]
format = "grouped"
max_line_length = 120
show_source = true
docstring-convention = "google"
class_attributes_order = ["field", "meta_class", "nested_class", "magic_method", "property_method", "static_method", "private_method", "method", "class_method"]
extended_default_ignore = []
inline-quotes = '"'
rst-roles = ["class", "method", "ref", "attr"]

[tool.flakeheaven.plugins]
pyflakes = ["+*"]
pycodestyle = ["+*", "-E203", "-E501"]
mccabe = ["+*"]
flake8-quotes = ["+*"]
"flake8-*" = ["+*"]
flake8-bandit = ["+*", "-S101"]
flake8-docstrings = ["+*", "-D100", "-D104", "-D105", "-D107", "-D200", "-D204", "-D205", "-D301", "-D400", "-D401", "-D402", "-D412", "-D413"]
flake8-rst-docstrings = ["+*", "-RST201", "-RST203", "-RST301", "-RST210", "-RST213"]
flake8-annotations = ["+*", "-ANN002", "-ANN003", "-ANN101", "-ANN102", "-ANN401"]
pep8-naming = ["+*", "-N812"]
flake8-darglint = ["+*"]

[tool.flakeheaven.exceptions."tests/"]
flake8-bandit = ["-S101"]

[tool.isort]
profile = "black"
combine_as_imports = true
balanced_wrapping = true
lines_after_imports = 2
include_trailing_comma = true
multi_line_output = 5
line_length = 120
known_third_party = ["invoke"]

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = true
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
fail-under = 90
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true

[tool.pytest]
norecursedirs = ["*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git __pycache__"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "abc/__version__.py",
    "pyproject.toml:version"
]
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
